Personal Assistant.ipynb
This is a multi-layer perceptron model. In this model I'm using bath normalization and also, everything is implemented from scratch, this is the first time I dive deeply into batch norm and also how to better initialize the weights and the biases of the neural network.  By the way, this is version one.